+++ changelog.md
@@ -2,8 +2,13 @@
 - Maybe I was wrong in my readme statement...
 - Plugin loader
 - Channels for plugin && app communication
-
+pre3:
+- Plugin can now register commands
+- !plugin list command
+- get plug by id
 ## TODO
+- dotfiles
 - Editor events
 - Better plugin api overall...
 - Frevi dotfiles and plugin autoload
+- mb multithreading?